Added comments to some components and separated db calls into controller files
