+++ server/controllers/gymController.js
@@ -12,38 +12,33 @@ module.exports = {
     },
 
     getUsersGyms: (req,res)=>{
-        req.db.users_gyms.find({user_id: req.user.id})
-            .then(gyms=>{
-                res.send(gyms)
-            })
-            .catch(err=>{
-                console.log(err)
-            })
+        req.db.GET_USERS_GYMS([req.user.id])
+        .then(gyms=>{
+            res.send(gyms)
+        })
+        .catch(err=>{
+            console.log(err)
+        })
     },
 
     postGym: (req, res)=>{
         req.db.gym_listings.insert(req.body)
-            .then(gym_listing =>{
-                req.db.users_gyms.insert({gym_id: gym_listing.id, user_id: req.user.id})
-                    .then(gym=>{
-                        res.send(gym)
-                    })
-                    .catch(err=>{
-                        console.log(err)
-                    })
-            })
-            .catch(err=>{
-                console.log(err)
-            })
+        .then(gym_listing =>{
+            res.send(gym_listing)
+        })
+        .catch(err=>{
+            console.log(err)
+        })
     },
 
-    postUsersGym: (req,res)=>{
-        req.db.users_gyms.insert({gym_id: req.body.gym_id, user_id: req.user.id})
-            .then(gym=>{
-                res.send(gym)
-            })
-            .catch(err=>{
-                console.log(err)
-            })
+    postUsersGym: async (req,res)=>{
+        try{
+            await  req.db.users_gyms.insert({gym_id: req.body.gym_id, user_id: req.user.id})
+            const gyms = await req.db.GET_USERS_GYMS([req.user.id])
+            res.send(gyms)
+        } catch (err){
+            console.log(err)
+            res.status(500).send(err)
+        }
     }
 }
