+++ server/controllers/tickController.js
@@ -2,7 +2,7 @@
 
 module.exports = {
     getTicks: (req, res)=>{
-        req.db.ticks.find({user_id: req.user.id})
+        req.db.GET_TICKS([req.user.id])
            .then(ticks=>{
                res.send(ticks)
            })
@@ -10,55 +10,33 @@ module.exports = {
                console.log(err)
            })
     },
-
-    postTick: (req, res)=>{
+ 
+    postTick: async (req, res)=>{
         let date = new Date()
-        req.db.ticks.insert({date_created: date, user_id: req.user.id, route_id: req.body.route_id})
-           .then(tick=>{
-               req.db.star_ratings.insert({date_created: date, user_id: req.user.id, route_id: req.body.route_id, stars: req.body.stars})
-                   .then(stars=>{
-                       req.db.difficulty_ratings.insert({date_created: date, user_id: req.user.id, route_id: req.body.route_id, rating: req.body.rating})
-                           .then(rating=>{
-                               req.db.todos.destroy({user_id: req.user.id, route_id: req.body.route_id})
-                                   .then(todo=>{
-                                       res.send({tick: tick, stars: stars, rating: rating, todoRemoved: todo})
-                                   })
-                                   .catch(err=>{
-                                       console.log(err)
-                                   })
-                           })
-                           .catch(err=>{
-                               console.log(err)
-                           })
-                   })
-                   .catch(err=>{
-                       console.log(err)
-                   })
-           })
-           .catch(err=>{
-               console.log(err)
-           })
+        try { 
+            await req.db.ticks.insert({date_created: date, user_id: req.user.id, route_id: req.body.route_id})
+            await req.db.star_ratings.insert({date_created: date, user_id: req.user.id, route_id: req.body.route_id, stars: req.body.stars})
+            await req.db.difficulty_ratings.insert({date_created: date, user_id: req.user.id, route_id: req.body.route_id, rating: req.body.rating})
+            const ticks = await req.db.GET_TICKS([req.user.id])
+            res.send(ticks)
+        } catch(err){
+            console.log(err)
+            res.status(500).send(err)
+        }
     },
 
-    deleteTick: (req, res)=>{
-        req.db.ticks.destroy({user_id: 3, route_id: req.params.route_id})
-            .then(tick=>{
-                req.db.star_ratings.destroy({user_id: req.user.id, route_id: req.params.route_id})
-                    .then(star=>{
-                        req.db.difficulty_ratings.destroy({user_id: req.user.id, route_id: req.params.route_id})
-                            .then(rating=>{
-                                res.send({message: 'tick removed', tick: tick[0], star: star[0], rating: rating[0]})
-                            })
-                            .catch(err=>{
-                                console.log(err)
-                            })
-                    })
-                    .catch(err=>{
-                        console.log(err)
-                    })
-            })
-            .catch(err=>{
-                console.log(err)
-            })
-     }
+    deleteTick: async (req, res)=>{
+        try { 
+            await req.db.ticks.destroy({user_id: 3, route_id: req.params.route_id})
+            await req.db.star_ratings.destroy({user_id: req.user.id, route_id: req.params.route_id})
+            await req.db.difficulty_ratings.destroy({user_id: req.user.id, route_id: req.params.route_id})
+            const ticks = await req.db.GET_TICKS([req.user.id])
+            res.send({message: 'tick removed', ticks: ticks})
+        } catch(err){
+            console.log(err)
+            res.status(500).send(err)
+        }
+    },
+
+
 }
