+++ src/components/addGym/addGym.css
@@ -1,11 +1,4 @@
 
-.popup-content h3{
-    align-self: center;
-    margin-top: 3px;
-    margin-bottom: 2px;
-    font-size: 1.6em;
-    color:#333640;
-}
 .postGym-button{
     margin-top: 30px;
     padding: 10px;
