--- src/components/commentBox/CommentBox.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-
-
-class CommentBox extends Component {
-    render() {
-      return (
-        <div>
-            CommentBox
-        </div>
-      );
-    }
-  }
-  
-  export default CommentBox;
