+++ src/components/editRoute/EditRoute.js
@@ -1,16 +1,16 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
+import grades from '../../helperFiles/grades'
+import colors from '../../helperFiles/colors'
 import getMuiTheme        from 'material-ui/styles/getMuiTheme';
 import MuiThemeProvider   from 'material-ui/styles/MuiThemeProvider';
 import TextField from 'material-ui/TextField';
 import SelectField from 'material-ui/SelectField';
 import MenuItem from 'material-ui/MenuItem';
 import DatePicker from 'material-ui/DatePicker';
-import './editRoute.css';
-import FineUploaderS3 from 'fine-uploader-wrappers/s3';
 import FileInput from '../fileInput/FileInput';
-// import Gallery from 'react-fine-uploader'
-// import 'react-fine-uploader/gallery/gallery.css'
+import './editRoute.css';
+
 
 class EditRoute extends Component {
   constructor(){
@@ -28,6 +28,11 @@ class EditRoute extends Component {
           disabled: false,
       }
   }
+
+  //if this component is rendered from the Edit button on RouteView we will receive a routeInfo prop
+  //if we receive this prop we want to populate state with it so we can see and edit the info with the UI
+  //else the component is loaded from the Add Route button on the dashboard and we just want to set the 
+  //dates, selected gym, and setter
   componentWillMount(){
     if(this.props.routeInfo){
       this.setState({
@@ -45,6 +50,8 @@ class EditRoute extends Component {
       })
     }
   }
+
+  //material UI is funcky with its onChanges so I needed a separate handler for each
   selectValue(event, i, value, name){
     this.setState({
         [name]: value
@@ -65,6 +72,12 @@ class EditRoute extends Component {
         removal_date: date
       })
   }
+
+  //I wanted the freedom to know exactly what I was sending to the db so I set route in the postRoute function
+  //then, if we are coming from Edit Route I add the route id... coming from Add Route will not have an ID
+  // the updateRoute function is called from either the Dashboard or RouteView depending on 
+  //where the component was mounted from it will call an axios.put or axios.post depending
+  //then clear state so the component will be ready for the next render
   postRoute(){
     let route = {
       gym_id: this.props.gyms.selectedGym,
@@ -99,67 +112,6 @@ class EditRoute extends Component {
   }
 
     render() {
-      let boulderingGrades = [
-        <MenuItem key={1} value='v0' primaryText="V0" />,
-        <MenuItem key={2} value='v1' primaryText="V1" />,
-        <MenuItem key={3} value='v2' primaryText="V2" />,
-        <MenuItem key={4} value='v3' primaryText="V3" />,
-        <MenuItem key={5} value='v4' primaryText="V4" />,
-        <MenuItem key={6} value='v5' primaryText="V5" />,
-        <MenuItem key={7} value='v6' primaryText="V6" />,
-        <MenuItem key={8} value='v7' primaryText="V7" />,
-        <MenuItem key={9} value='v8' primaryText="V8" />,
-        <MenuItem key={10} value='v9' primaryText="V9" />,
-        <MenuItem key={11} value='v10' primaryText="V10" />,
-        <MenuItem key={12} value='v11' primaryText="V11" />,
-        <MenuItem key={13} value='v12' primaryText="V12" />,
-        <MenuItem key={14} value='v13' primaryText="V13" /> 
-      ]
-      let sportGrades = [
-        <MenuItem key={15} value='5.8' primaryText="5.8" />,
-        <MenuItem key={16} value='5.9' primaryText="5.9" />,
-        <MenuItem key={17} value='5.10a' primaryText="5.10a" />,
-        <MenuItem key={18} value='5.10b' primaryText="5.10b" />,
-        <MenuItem key={19} value='5.10c' primaryText="5.10c" />,
-        <MenuItem key={20} value='5.10d' primaryText="5.10d" />,
-        <MenuItem key={21} value='5.11a' primaryText="5.11a" />,
-        <MenuItem key={22} value='5.11b' primaryText="5.11b" />,
-        <MenuItem key={23} value='5.11c' primaryText="5.11c" />,
-        <MenuItem key={24} value='5.11d' primaryText="5.11d" />,
-        <MenuItem key={25} value='5.12a' primaryText="5.12a" />,
-        <MenuItem key={26} value='5.12b' primaryText="5.12b" />,
-        <MenuItem key={27} value='5.12c' primaryText="5.12c" />,
-        <MenuItem key={28} value='5.12d' primaryText="5.12d" />,
-        <MenuItem key={29} value='5.13a' primaryText="5.13a" />,
-        <MenuItem key={30} value='5.13b' primaryText="5.13b" />,
-        <MenuItem key={31} value='5.13c' primaryText="5.13c" />,
-        <MenuItem key={32} value='5.13d' primaryText="5.13d" />,
-        <MenuItem key={33} value='5.14' primaryText="5.14" /> 
-      ]
-      let colors = [
-        <MenuItem key={34} value='blue' primaryText="" style={{backgroundColor:'blue'}}/>,
-        <MenuItem key={35} value='red' primaryText="" style={{backgroundColor:'red'}}/>,
-        <MenuItem key={36} value='green' primaryText="" style={{backgroundColor:'green'}}/>,
-        <MenuItem key={37} value='white' primaryText="" style={{backgroundColor:'white'}}/>,
-        <MenuItem key={38} value='purple' primaryText="" style={{backgroundColor:'purple'}}/>,
-        <MenuItem key={39} value='black' primaryText="" style={{backgroundColor:'black'}}/>,
-        <MenuItem key={40} value='tan' primaryText="" style={{backgroundColor:'tan'}}/>,
-        <MenuItem key={41} value='pink' primaryText="" style={{backgroundColor:'pink'}}/>,
-        <MenuItem key={42} value='grey' primaryText="" style={{backgroundColor:'grey'}}/>,
-        <MenuItem key={43} value='brown' primaryText="" style={{backgroundColor:'brown'}}/>,
-        <MenuItem key={44} value='yellow' primaryText="" style={{backgroundColor:'yellow'}}/>,
-      ]
-      const uploader = new FineUploaderS3({
-        options: {
-            request: {
-                endpoint: "http://fineuploadertest.s3.amazonaws.com",
-                accessKey: "AKIAIXVR6TANOGNBGANQ"
-            },
-            signature: {
-                endpoint: "/vendor/fineuploader/php-s3-server/endpoint.php"
-            },
-        } 
-    })
       return (
         <div className='popup-background' style={{visibility: this.props.visibility}}>
             <div className='popup-content'> 
@@ -168,7 +120,6 @@ class EditRoute extends Component {
               :
                 <span className='title-container'>  <h3> Add Route </h3> <button className='exit-popup-button' onClick={()=>this.props.AddRouteVisibility('hidden')}> X</button></span>
             }
-          
             <MuiThemeProvider muiTheme={getMuiTheme() }>
               <div className='editRouteForm'>
                   <div className='firstRow'>
@@ -209,11 +160,10 @@ class EditRoute extends Component {
                           onChange={(event, i, value, name)=>this.selectValue(event, i, value, 'difficulty')}>
                           {
                             this.state.type === 'bouldering' ?
-                            boulderingGrades
+                            grades.boulderingGrades
                             :
-                            sportGrades
+                            grades.sportGrades
                           }
-                        
                       </SelectField>
                   </div>
                   <div className='secondRow'>
@@ -248,18 +198,11 @@ class EditRoute extends Component {
                           />
                       </div>
                       <div className='secondColumn'>
-                        {/* <Gallery 
-                            uploader={ uploader }
-                            multiple={false}
-                            fileInput-multiple={false}
-                            dropzone-disabled={ true }
-                            /> */}
                         <FileInput route_id={this.state.id} />
                       </div>
                   </div>
               </div>
             </MuiThemeProvider>
-     
             <button className='green-button editRouteDone-button' onClick={()=>this.postRoute()}>Done</button>
             </div>
         </div>
