+++ src/components/editRoute/editRoute.css
@@ -13,9 +13,5 @@
     align-items: center;
 }
 .secondColumn{
-    /* border: 1px solid #333640;
-    height: 140px;
-    width: 120px;
-    margin: 5px;*/
     margin-right: 10px; 
 }
