+++ src/components/fileInput/FileInput.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import {addImage, removeImage} from '../../redux/action-creators'
-import axios from 'axios'
 import './fileInput.css'
 
 
@@ -15,6 +14,9 @@ class FileInput extends Component {
       this.handleSubmit = this.handleSubmit.bind(this);
     }
 
+//receives the route image from props from the EditRoute component. 
+//if it exists.... is not null then set it to state and uploaded to true
+//this effects which button will show up
     componentWillMount(){
         if(this.props.routeImage){
             this.setState({
@@ -24,25 +26,24 @@ class FileInput extends Component {
         }
     }
 
-/// handles form submit
+//this is called with +pic is clicked... it renders the preview
     handleSubmit(event) {
       event.preventDefault();
-      // give us the object for handling form-data aka files from our computer
+    // give us the object for handling form-data aka files from our computer
        const reader = new FileReader();
-    // function that reads the file and converts it to base64.
+    // function .onload reads the file and .result converts it to base64.
        const readImage = (file) =>{
-        reader.readAsDataURL(file);
-        const allowedFileTypes = ['png', 'jpg', 'jpeg', 'gif' ]
-        const fileType = file.name.split('.')[1];
-        if(allowedFileTypes.includes(fileType)){
-            console.log(file)
-            reader.onload = function(){
-                setImage(reader.result, file);
-               }
-           }else{
-                alert("we only except images");
-            }
-        }
+                reader.readAsDataURL(file);
+                const allowedFileTypes = ['png', 'jpg', 'jpeg', 'gif' ]
+                const fileType = file.name.split('.')[1];
+                if(allowedFileTypes.includes(fileType)){
+                    reader.onload = function(){
+                        setImage(reader.result, file);
+                    }
+                }else{
+                        alert("we only except images");
+                    }
+                }
     // Sets base64 to state.
        let setImage = (base64, file) => {
         this.setState({
@@ -54,11 +55,12 @@ class FileInput extends Component {
        setImage = setImage.bind(this);
 
     //invokes readImage and passed in the file object that we get from the form.
-       readImage(this.fileInput.files[0])
-       
+       readImage(this.fileInput.files[0]) 
     }
 
-    uploadImage(){
+//sends the file through redux to send to the db and set to redux state
+//I needed to set the image to local state here as well to allow the preview to function
+uploadImage(){
         let formData = new FormData();
         formData.set('abc', this.state.fileToUpload)
         Promise.resolve(this.props.addImage(this.props.route_id, formData))
@@ -69,9 +71,10 @@ class FileInput extends Component {
                     uploaded: true
                 })
             })
-    }
+}
 
-    removeImage(){
+//gets the filename from state and sends it through redux to remove the image from our db / s3/ and redux state
+removeImage(){
         let fileName = this.state.image.split('.com/')[1]
         Promise.resolve(this.props.removeImage(this.props.route_id, fileName))
             .then(response=>{
@@ -81,36 +84,34 @@ class FileInput extends Component {
                     uploaded:false
                 })
             })
-    }
+}
     
     render() {
       return (
             <div className='routeImage' style={this.state.image ? {backgroundImage:`url(${this.state.image})`}: {fontSize: '.8em'}}>
-
-            {this.state.image ?
-                <div>
-                    { this.state.uploaded ?
-                        <button className='pink-button remove-image' onClick={()=>this.removeImage()}>X</button>
-                        :
-                        <div>
-                            <button className='inputfile-upload' onClick={()=>this.uploadImage()}></button>
+                {this.state.image ?
+                    <div>
+                        { this.state.uploaded ?
                             <button className='pink-button remove-image' onClick={()=>this.removeImage()}>X</button>
-                        </div>
-                    }
-                </div>
-            :
-            <div>
-                    <input className="inputfile"
-                    name='file'
-                    id='file'
-                    type="file"
-                    ref={input => {this.fileInput = input}}
-                    onChange={this.handleSubmit}
-                    />
-                    <div className='inputfile-label green-button'><label className='label' htmlFor="file" >+ pic</label></div>
-            </div>
-            
-            }   
+                            :
+                            <div>
+                                <button className='inputfile-upload' onClick={()=>this.uploadImage()}></button>
+                                <button className='pink-button remove-image' onClick={()=>this.removeImage()}>X</button>
+                            </div>
+                        }
+                    </div>
+                :
+                    <div>
+                            <input className="inputfile"
+                            name='file'
+                            id='file'
+                            type="file"
+                            ref={input => {this.fileInput = input}}
+                            onChange={this.handleSubmit}
+                            />
+                            <div className='inputfile-label green-button'><label className='label' htmlFor="file" >+ pic</label></div>
+                    </div>
+                }   
             </div>
       );
     }
