+++ src/components/gyms/gyms.css
@@ -13,10 +13,3 @@
     padding: 5px;
     font-size: .85em;
 }
-.title-container{
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-    justify-content: space-between;
-    align-items: center;
-}
