+++ src/components/header/Header.js
@@ -11,15 +11,19 @@ class Header extends Component {
       gymName: 'Select a Gym'
     }
   }
+  //takes the selected gym from redux (an id int) and finds the gym name 
+  //from the usersGyms to set to gymName on local state
   componentWillMount(){
-    this.props.gyms.usersGyms.map(gym=>{
+    this.props.gyms.usersGyms.map(gym=>{ 
       if(gym.gym_id === this.props.gyms.selectedGym){
-        this.setState({
+       this.setState({
           gymName: gym.name
         })
       }
     })
   }
+  //Since the selected gym can change outside of this component I need to 
+  //inform the component to listen for changes in its props
   componentWillReceiveProps(props){
     props.gyms.usersGyms.map(gym=>{
       if(gym.gym_id === props.gyms.selectedGym){
@@ -46,14 +50,13 @@ class Header extends Component {
              </div>
           </Link>
               {this.props.user.role === 'setter' ?
-              <select value={this.props.user.temporaryRole} name='temporaryRole' className='roleSelect' onChange={(event)=>this.props.changeRole(event.target.value)}>
-                <option value='setter'>Setter</option>
-                <option value='climber'>Climber</option>
-              </select>
+                <select value={this.props.user.temporaryRole} name='temporaryRole' className='roleSelect' onChange={(event)=>this.props.changeRole(event.target.value)}>
+                  <option value='setter'>Setter</option>
+                  <option value='climber'>Climber</option>
+                </select>
               : 
-              <p className='roleSelect'>Climber</p>
+                <p className='roleSelect'>Climber</p>
               }
-     
         </div>
        </div>
       );
