+++ src/components/header/header.css
@@ -23,9 +23,7 @@
     height: 40px;
     width: 50px;
     margin-left: 10px;
-
 }
-
 .header-logo:hover{
     -moz-transform: scaleX(-1);
     -o-transform: scaleX(-1);
@@ -34,7 +32,6 @@
     filter: FlipH;
     -ms-filter: "FlipH";
 }
-
 .userImage-select-div{
     display: flex;
     flex-direction: column;
@@ -64,7 +61,6 @@
     height: 60px;
     margin-top:10px;
 }
-
 .appTitle-select{
     margin-top: 0px;
     font-size: 1.2em;
