+++ src/components/routes/routes.css
@@ -1,4 +1,3 @@
-
 .route-box{
     display: flex;
     flex-direction: row;
