+++ src/helperFiles/colors.js
@@ -1,0 +1,19 @@
+import React, { Component } from 'react';
+import MenuItem from 'material-ui/MenuItem';
+
+
+const colors = [
+    <MenuItem key={34} value='blue' primaryText="" style={{backgroundColor:'blue'}}/>,
+    <MenuItem key={35} value='red' primaryText="" style={{backgroundColor:'red'}}/>,
+    <MenuItem key={36} value='green' primaryText="" style={{backgroundColor:'green'}}/>,
+    <MenuItem key={37} value='white' primaryText="" style={{backgroundColor:'white'}}/>,
+    <MenuItem key={38} value='purple' primaryText="" style={{backgroundColor:'purple'}}/>,
+    <MenuItem key={39} value='black' primaryText="" style={{backgroundColor:'black'}}/>,
+    <MenuItem key={40} value='tan' primaryText="" style={{backgroundColor:'tan'}}/>,
+    <MenuItem key={41} value='pink' primaryText="" style={{backgroundColor:'pink'}}/>,
+    <MenuItem key={42} value='grey' primaryText="" style={{backgroundColor:'grey'}}/>,
+    <MenuItem key={43} value='brown' primaryText="" style={{backgroundColor:'brown'}}/>,
+    <MenuItem key={44} value='yellow' primaryText="" style={{backgroundColor:'yellow'}}/>,
+  ]
+
+export default colors
